requirements.txt
python-telegram-bot[webhooks]==20.6
chromadb==0.5.0
InstructorEmbedding
requests
numpy<2.0.0

